Support current session metrics
For this to happen, a rewrite of the metrics subsystem was needed. It
now seamlessly supports general metrics, per session metrics as well as
cumulative metrics. For this to work, a type instance change was
required, rendering this into a breaking API change as all RRDs need to
be renamed