Switch to transmission-rpc
transmissionrpc as a python package is obviously unmaintained.
A fork that went until mid 2020, transmissionrpc-ng was tried out, but

a) it's already 3 years that it has seen 0 movement
b) it decided to use pycurl, which is a native extension, making testing difficult

There is another fork, which although not explicitly stated, contains
the original one's copyright notices. We switch to using that one. For
now we stick to versions below 3.0 as there appears to have been a
breakage in 3.0

Signed-off-by: Alexandros Kosiaris <akosiaris@gmail.com>