Fix metric breakage added in transmission-rpc 4.0.0
Why:
In transmission-rpc 4.0.0 and later, metric names were altered to be
more Pythonic. However, that breaks history for installation of this
plugin. We want to see if we can keep them functional

What:
Become more transmission-rpc aware, by importing specific objects we use
instead of the module. Based on the successful or failed import of
SessionStats, a class introduced in 4.0.0, codify different versions the
metric names and use a name key to keep the names constant

Update tests to reflect the above

Signed-off-by: Alexandros Kosiaris <akosiaris@gmail.com>